+++ packages/adminui-vue/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [1.6.4]
 ### Fixed
  - "Always use domain for login" checkbox should be visible in navbar User menu in case "Negotiate" auth is enabled for app [UBDF-9586]
+ - `class UForm` fixed the problem when two identical forms have a common store state and replace each other's data
+ - `class UForm` Now any method or state property of the UForm store can be overwritten by local form store
 
 ## [1.6.3]
 ### Fixed